Move util function to separate module to resolve circular import
